# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Revenge Media Server API
  version: 0.1.0
  description: |
    Modern, self-hosted media server with Go backend and SvelteKit frontend.

    This is the skeleton API for v0.1.0 - only health endpoints are implemented.
  contact:
    name: Revenge Development Team
    url: https://github.com/lusoris/revenge
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.revenge.example.com
    description: Production server
...
paths:
...
  /api/v1/settings/server:
    get:
      summary: List all server settings
      description: Retrieve all server-wide configuration settings
      operationId: listServerSettings
      tags:
        - settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of server settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServerSetting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/settings/server/{key}:
    get:
      summary: Get a server setting
      description: Retrieve a specific server setting by key
      operationId: getServerSetting
      tags:
        - settings
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Setting key
          example: server.name
      responses:
        '200':
          description: Server setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerSetting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'

    put:
      summary: Update a server setting
      description: Update the value of a server setting
      operationId: updateServerSetting
      tags:
        - settings
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Setting key
          example: server.name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingValue'
      responses:
        '200':
          description: Updated server setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerSetting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/settings/user:
    get:
      summary: List current user's settings
      description: Retrieve all settings for the authenticated user
      operationId: listUserSettings
      tags:
        - settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSetting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/settings/user/{key}:
    get:
      summary: Get a user setting
      description: Retrieve a specific user setting by key
      operationId: getUserSetting
      tags:
        - settings
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Setting key
          example: theme
      responses:
        '200':
          description: User setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSetting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'

    put:
      summary: Update a user setting
      description: Update the value of a user setting
      operationId: updateUserSetting
      tags:
        - settings
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Setting key
          example: theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingValue'
      responses:
        '200':
          description: Updated user setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSetting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'

    delete:
      summary: Delete a user setting
      description: Delete a user setting (revert to default)
      operationId: deleteUserSetting
      tags:
        - settings
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Setting key
          example: theme
      responses:
        '204':
          description: Setting deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'

  # User endpoints
  /api/v1/users/me:
    get:
      summary: Get current user
      description: Get the currently authenticated user's profile
      operationId: getCurrentUser
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'

    put:
      summary: Update current user
      description: Update the authenticated user's profile
      operationId: updateCurrentUser
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/users/me/preferences:
    get:
      summary: Get user preferences
      description: Get notification and display preferences for the authenticated user
      operationId: getUserPreferences
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'

    put:
      summary: Update user preferences
      description: Update notification and display preferences
      operationId: updateUserPreferences
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdate'
      responses:
        '200':
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/users/me/avatar:
    post:
      summary: Upload avatar
      description: Upload a new avatar image
      operationId: uploadAvatar
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Avatar image file (JPEG, PNG, GIF, WebP, max 5MB)
      responses:
        '200':
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Avatar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/users/{userId}:
    get:
      summary: Get user by ID
      description: Get a user's public profile information
      operationId: getUserById
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
components:
  schemas:
    ServerSetting:
      type: object
      required:
        - key
        - value
        - data_type
      properties:
        key:
          type: string
          description: Setting key
          example: server.name
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
          description: Setting value (type varies based on data_type)
          example: "My Revenge Server"
        description:
          type: string
          description: Human-readable description
          example: Server display name
        category:
          type: string
          description: Setting category for organization
          example: general
        data_type:
          type: string
          description: Expected data type
          enum: [string, integer, boolean, float, json]
          example: string
        is_secret:
          type: boolean
          description: Whether the setting contains sensitive data
          example: false
        is_public:
          type: boolean
          description: Whether the setting is exposed in public API
          example: true
    UserSetting:
      type: object
      required:
        - user_id
        - key
        - value
        - data_type
      properties:
        user_id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        key:
          type: string
          description: Setting key
          example: theme
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
          description: Setting value (type varies based on data_type)
          example: "dark"
        description:
          type: string
          description: Human-readable description
          example: UI theme preference
        category:
          type: string
          description: Setting category for organization
          example: ui
        data_type:
          type: string
          description: Expected data type
          enum: [string, integer, boolean, float, json]
          example: string
    SettingValue:
      type: object
      required:
        - value
      properties:
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
          description: The new value for the setting
          example: "dark"
    User:
      type: object
      required:
        - id
        - username
        - email
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Unique username
          example: johndoe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
        display_name:
          type: string
          description: Display name
          example: John Doe
        avatar_url:
          type: string
          description: Avatar image URL
          example: /avatars/550e8400-e29b-41d4-a716-446655440000/avatar.jpg
        locale:
          type: string
          description: Locale preference
          example: en-US
        timezone:
          type: string
          description: Timezone
          example: America/New_York
        qar_enabled:
          type: boolean
          description: QAR (adult content) access enabled
          example: false
        is_active:
          type: boolean
          description: Account is active
          example: true
        is_admin:
          type: boolean
          description: Has admin privileges
          example: false
        email_verified:
          type: boolean
          description: Email has been verified
          example: true
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2024-01-01T00:00:00Z
        last_login_at:
          type: string
          format: date-time
          description: Last login timestamp
          example: 2024-01-15T12:30:00Z
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/v1/auth/login
security: []